adding only not null items to the row
public class FastTableDataBuilder

object value = ConvertValueToTargetType(valueAsString);

if(value is not null)
{
    newRowData[column] = value;
}

Test results for old code:
Memory increase in bytes: 107,673,200
Time of loading the CSV was 00:00:00.2559820.
Time of reading the CSV was 00:00:00.2945257.

Test results for new code:
Memory increase in bytes: 12,026,864
Time of loading the CSV was 00:00:00.0466975.
Time of reading the CSV was 00:00:00.1562671.